[refactor] 기존 Entity 패키지 이동 (#341)
