+++ src/main/java/org/sopt/app/application/poke/PokeHistoryService.java
@@ -9,7 +9,7 @@ import lombok.val;
 import org.sopt.app.application.poke.PokeInfo.PokeHistoryInfo;
 import org.sopt.app.common.exception.BadRequestException;
 import org.sopt.app.common.response.ErrorCode;
-import org.sopt.app.domain.entity.PokeHistory;
+import org.sopt.app.domain.entity.poke.PokeHistory;
 import org.sopt.app.interfaces.postgres.PokeHistoryRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
