+++ src/main/java/org/sopt/app/application/poke/PokeMessageService.java
@@ -6,7 +6,7 @@ import java.util.Random;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
-import org.sopt.app.domain.entity.PokeMessage;
+import org.sopt.app.domain.entity.poke.PokeMessage;
 import org.sopt.app.domain.enums.PokeMessageType;
 import org.sopt.app.interfaces.postgres.PokeMessageRepository;
 import org.springframework.stereotype.Service;
