+++ src/main/java/org/sopt/app/application/soptamp/SoptampPointService.java
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.sopt.app.application.soptamp.SoptampPointInfo.PartPoint;
 import org.sopt.app.application.soptamp.SoptampPointInfo.Point;
-import org.sopt.app.domain.entity.SoptampPoint;
-import org.sopt.app.domain.entity.SoptampUser;
+import org.sopt.app.domain.entity.soptamp.SoptampPoint;
+import org.sopt.app.domain.entity.soptamp.SoptampUser;
 import org.sopt.app.domain.enums.Part;
 import org.sopt.app.domain.enums.UserStatus;
 import org.sopt.app.interfaces.postgres.soptamp_point.SoptampPointRepository;
