+++ src/main/java/org/sopt/app/interfaces/postgres/PokeHistoryRepository.java
@@ -1,7 +1,7 @@
 package org.sopt.app.interfaces.postgres;
 
 import java.util.List;
-import org.sopt.app.domain.entity.PokeHistory;
+import org.sopt.app.domain.entity.poke.PokeHistory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
