+++ src/main/java/org/sopt/app/interfaces/postgres/PokeMessageRepository.java
@@ -1,10 +1,8 @@
 package org.sopt.app.interfaces.postgres;
 
-import org.sopt.app.domain.entity.PokeMessage;
+import org.sopt.app.domain.entity.poke.PokeMessage;
 import org.sopt.app.domain.enums.PokeMessageType;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
