+++ src/main/java/org/sopt/app/interfaces/postgres/StampRepository.java
@@ -2,7 +2,7 @@ package org.sopt.app.interfaces.postgres;
 
 import java.util.List;
 import java.util.Optional;
-import org.sopt.app.domain.entity.Stamp;
+import org.sopt.app.domain.entity.soptamp.Stamp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StampRepository extends JpaRepository<Stamp, Long> {
