+++ src/test/java/org/sopt/app/application/MissionServiceTest.java
@@ -17,8 +17,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.sopt.app.application.mission.MissionInfo.Completeness;
 import org.sopt.app.application.mission.MissionInfo.Level;
 import org.sopt.app.application.mission.MissionService;
-import org.sopt.app.domain.entity.Mission;
-import org.sopt.app.domain.entity.Stamp;
+import org.sopt.app.domain.entity.soptamp.Mission;
+import org.sopt.app.domain.entity.soptamp.Stamp;
 import org.sopt.app.interfaces.postgres.MissionRepository;
 import org.sopt.app.interfaces.postgres.StampRepository;
 import org.sopt.app.presentation.mission.MissionRequest.RegisterMissionRequest;
