+++ src/test/java/org/sopt/app/application/SoptampPointServiceTest.java
@@ -18,7 +18,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.sopt.app.application.soptamp.SoptampPointInfo.Point;
 import org.sopt.app.application.soptamp.SoptampPointService;
-import org.sopt.app.domain.entity.SoptampPoint;
+import org.sopt.app.domain.entity.soptamp.SoptampPoint;
 import org.sopt.app.domain.enums.UserStatus;
 import org.sopt.app.interfaces.postgres.soptamp_point.SoptampPointRepository;
 import org.springframework.test.util.ReflectionTestUtils;
