+++ src/test/java/org/sopt/app/application/SoptampUserServiceTest.java
@@ -19,7 +19,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.sopt.app.application.soptamp.SoptampUserInfo;
 import org.sopt.app.application.soptamp.SoptampUserService;
 import org.sopt.app.common.exception.BadRequestException;
-import org.sopt.app.domain.entity.SoptampUser;
+import org.sopt.app.domain.entity.soptamp.SoptampUser;
 import org.sopt.app.interfaces.postgres.SoptampUserRepository;
 
 @ExtendWith(MockitoExtension.class)
