+++ src/test/java/org/sopt/app/facade/PokeFacadeTest.java
@@ -48,8 +48,8 @@ import org.sopt.app.application.user.UserService;
 import org.sopt.app.common.fixtures.PokeFixture;
 import org.sopt.app.common.fixtures.UserFixture;
 import org.sopt.app.domain.entity.Friend;
-import org.sopt.app.domain.entity.PokeHistory;
-import org.sopt.app.domain.entity.PokeMessage;
+import org.sopt.app.domain.entity.poke.PokeHistory;
+import org.sopt.app.domain.entity.poke.PokeMessage;
 import org.sopt.app.domain.entity.User;
 import org.sopt.app.domain.enums.FriendRecommendType;
 import org.sopt.app.domain.enums.Friendship;
